Migrate Functions into src/ folder for structure + update to .NET 8 isolated
