+++ .vscode/settings.json
@@ -1,0 +1,7 @@
+{
+  "azureFunctions.deploySubpath": "src\\DurableOrchestrator/bin/Release/net8.0/publish",
+  "azureFunctions.projectLanguage": "C#",
+  "azureFunctions.projectRuntime": "~4",
+  "debug.internalConsoleOptions": "neverOpen",
+  "azureFunctions.preDeployTask": "publish (functions)"
+}
