+++ DurableOrchestrator.sln
@@ -1,0 +1,25 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 17
+VisualStudioVersion = 17.5.002.0
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "DurableOrchestrator", "src\DurableOrchestrator\DurableOrchestrator.csproj", "{89CC8406-9DF5-44D9-98A7-6F9D794EB79A}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{89CC8406-9DF5-44D9-98A7-6F9D794EB79A}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{89CC8406-9DF5-44D9-98A7-6F9D794EB79A}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{89CC8406-9DF5-44D9-98A7-6F9D794EB79A}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{89CC8406-9DF5-44D9-98A7-6F9D794EB79A}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {B30DA6BE-C55F-4508-A862-527E6B591FA4}
+	EndGlobalSection
+EndGlobal
