--- Properties/launchSettings.json
@@ -1,9 +1,0 @@
-{
-  "profiles": {
-    "event_processing": {
-      "commandName": "Project",
-      "commandLineArgs": "--port 7063",
-      "launchBrowser": false
-    }
-  }
-}
